+++ backend/todo-list/src/main/java/com/example/todolist/controller/CardController.java
@@ -1,6 +1,7 @@
 package com.example.todolist.controller;
 
 
+import com.example.todolist.ApiResponse;
 import com.example.todolist.domain.Card;
 import com.example.todolist.domain.User;
 import com.example.todolist.repository.CardRepository;
@@ -13,6 +14,8 @@ import org.springframework.web.bind.annotation.*;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.util.HashMap;
+
+import static com.example.todolist.ApiResponse.Status.SUCCESS;
 
 @RestController
 @RequestMapping("/cards")
@@ -28,16 +31,15 @@ public class CardController {
     }
 
     public User getUserFromToken(HttpServletRequest request) {
-        System.out.println("getUserFromToken!!!");
         return (User) request.getAttribute("user");
     }
 
     @PostMapping
-    public ResponseEntity<Card> create(@RequestBody HashMap<String, String> cardInfo, HttpServletRequest request) {
+    public ResponseEntity<ApiResponse> create(@RequestBody HashMap<String, String> cardInfo, HttpServletRequest request) {
         User tokenUser = getUserFromToken(request);
         Card card = new Card(tokenUser, cardInfo.get("title"), cardInfo.get("contents"), cardInfo.get("status"));
         cardRepository.save(card);
-        return new ResponseEntity<>(card, HttpStatus.OK);
+        return new ResponseEntity<>(new ApiResponse(card, SUCCESS), HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
